+++ lib/vime/ui/spinner.rb
@@ -6,7 +6,7 @@ module Vime
   module Ui
     class Spinner < Component
       def call
-        content_tag "vm-spinner", process_attrs(dom_attrs)
+        content_tag "vm-spinner", nil, process_attrs(dom_attrs)
       end
     end
   end